New Jump feature

+++ Assets/Backend/Model/Ability/AbilityArgs/JumpAbilityArgs.cs
@@ -7,7 +7,7 @@ namespace Assets.Backend.Model
     [Serializable]
     public class JumpAbilityArgs
     {
-        public int JumpForce;
+        public float JumpForce;
         internal bool IsGrounded;
         internal int JumpsDone;
         public int JumpsMax;
